requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
scipy==1.11.4